pyproject.toml
[project]
name = "azure-devops-filler"
version = "0.1.0"
description = "CLI para preencher Tasks no Azure DevOps automaticamente"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "icalendar>=5.0",
]

[project.scripts]
adf = "azure_devops_filler.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "respx>=0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/azure_devops_filler"]

[tool.hatch.build.targets.sdist]
include = ["src/azure_devops_filler"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
